+++ ksmartteam01/src/main/resources/templates/admin/goods/goodsLargeCategoryInsert.html
@@ -90,9 +90,7 @@
 										<div class="ln_solid"></div>
 										<div class="item form-group">
 											<div class="col-md-6 col-sm-6 offset-md-3">
-												<button class="btn btn-primary" type="button">Cancel</button>
-												<button class="btn btn-primary" type="reset">Reset</button>
-												<button type="submit" class="btn btn-success">Submit</button>
+												<button type="submit" class="btn btn-success">등록</button>
 											</div>
 										</div>
 									</form>
