+++ ksmartteam01/src/main/resources/templates/admin/goods/goodsMediumCategoryList.html
@@ -40,7 +40,9 @@
                   <div class="row">
                    <div class="col-sm-12">
                     <div class="card-box table-responsive">
-                     <button class="buttonMoDel" type="button">등록</button>
+                      <div style="margin-right:10px;" class="d-flex justify-content-end">
+	                     <button type="button">등록</button>
+	                  </div>
                       <table id="datatable-checkbox" class="table table-striped table-bordered bulk_action" style="width:100%">
                        <thead>
                         <tr>
@@ -82,13 +84,12 @@
           </div>
          </div>
         </div>
+        </th:block>
+	<!-- end contents -->
 
-	
-			<!-- end footer -->
 	
 	<!-- 사용자 정의 JS 하단 -->
 	<th:block layout:fragment="customFootJs">		
-	
+		
 	</th:block>
-</body>
 </html>
