+++ ksmartteam01/src/main/resources/templates/admin/goods/goodsMediumCategoryModify.html
@@ -45,7 +45,7 @@
 						<div class="col-md-12 col-sm-12 ">
 							<div class="x_panel">
 								<div class="x_title">
-									<h2>중분류등록<small>different form elements</small></h2>
+									<h2>중분류수정<small></small></h2>
 									<ul class="nav navbar-right panel_toolbox">
 										<li><a class="collapse-link"><i class="fa fa-chevron-up"></i></a>
 										</li>
@@ -97,7 +97,7 @@
 										<div class="ln_solid"></div>
 										<div class="item form-group">
 											<div class="col-md-6 col-sm-6 offset-md-3">
-												<button type="submit" class="btn btn-success">등록</button>
+												<button type="submit" class="btn btn-success">수정</button>
 											</div>
 										</div>
 									</form>
