+++ ksmartteam01/src/main/resources/templates/admin/goods/goodsPaymentCancle.html
@@ -13,14 +13,14 @@
 		 <div class="">
 		  <div class="page-title">
 		    <div class="title_left">
-		     <h3>상품카테고리관리</h3>
+		     <h3>결제취소목록</h3>
 		    </div>
 		    <div class="clearfix"></div>
 			 <div class="row">
 			   <div class="col-md-12 col-sm-12 ">
                 <div class="x_panel">
                   <div class="x_title">
-                    <h2>대분류관리</h2>
+                    <h2></h2>
                     <ul class="nav navbar-right panel_toolbox">
                       <li><a class="collapse-link"><i class="fa fa-chevron-up"></i></a>
                       </li>
@@ -39,22 +39,18 @@
                  <div class="x_content">
                   <div class="row">
                    <div class="col-sm-12">
-                     <div class="card-box table-responsive">
-                      <div style="margin-right:10px;" class="d-flex justify-content-end">
-	                     <button type="button">등록</button>
-	                  </div>
+                    <div class="card-box table-responsive">
                       <table id="datatable-checkbox" class="table table-striped table-bordered bulk_action" style="width:100%">
-
                        <thead>
-	                 
                         <tr>
                           <th>
 							 <th><input type="checkbox" id="check-all" ></th>
 						  </th>
-						  <th>대분류코드</th>
-					      <th>대분류이름</th>
-						  <th>업데이트시각</th>
-					      <th>수정/삭제</th>
+                          <th>결제취소코드</th>
+                          <th>상품</th>
+                          <th>아이디</th>
+                          <th>취소일자</th>
+                          <th>취소사유</th>
                         </tr>
                       </thead>
                       <tbody>
@@ -63,14 +59,11 @@
 							 <th><input type="checkbox" id="check-all" ></th>
 						  </td>
                           <td>1</td>
-					      <td class="align-middle">사료</td>
-					      <td class="align-middle">
-					      	업데이트시각
-				          </td>
-					      <td class="align-middle">
-					      	 <button class="buttonMoDel" type="button">수정</button>
-		          			 <button class="buttonMoDel" type="button">삭제</button>
-					      </td> 
+                          <td>1</td>
+                          <td>1</td>
+                          <td>1</td>
+                          <td>1</td>
+
                         </tr>
                        </tbody>
                     </table>
@@ -92,4 +85,5 @@
 	<th:block layout:fragment="customFootJs">		
 	
 	</th:block>
+</body>
 </html>
